storage: enforce timestamp invariants in wal
