+++ src/storage/Cargo.toml
@@ -15,4 +15,5 @@ lazy_static = "1.4.0"
 log = "0.4.13"
 regex = "1.4.3"
 repr = { path = "../repr" }
+timely = { git = "https://github.com/TimelyDataflow/timely-dataflow", default-features = false, features = ["bincode"] }
 tokio = "1.2.0"
