+++ src/storage/src/compacter.rs
@@ -7,14 +7,49 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+//! Organize and maintain a compacted representation of persisted data.
+//!
+//! The `Compacter` task keeps track of the persisted updates for each persisted relation and
+//! periodically compacts that representation to use space proportional to the number of
+//! distinct rows in the relation. In order to do so, the `Compacter` maintains a `Trace` for
+//! each persisted relation. Note that the `Compacter` is currently a thread running on the
+//! Materialize process, but there's no conceptual reason it couldn't be a separate process, or
+//! even on a separate machine as long as it had access to the storage for the WAL (e.g. with a
+//! shared EBS volume).
+//!
+//! A `Trace` is basically a list of `Batch`s that represent a contiguous time interval, and a
+//! compaction frontier.
+//!
+//! A `Batch` is a consolidated list of updates that occured between times [lower, upper)
+//! where each update is of the form `(Row, time, diff)` and each `(Row, time)` pair occurs
+//! exactly once and all diffs are nonzero.
+//!
+//! Note that all `Batch`s keep their data on persistent storage. No data resides in memory
+//! (except currently we load all the data from batches into memory for compaction and on
+//!  restart but this will get fixed!).
+//!
+//! The `Coordinator` thread tells the `Compacter` when it needs to
+//!  * start keeping track of a new relation
+//!  * stop keeping track of a relation
+//!  * resume keeping track of a relation with an already initialized Trace (on restart)
+//!  * advance a relation's compaction frontier. Note that this doesn't automatically trigger
+//!    any actual compaction. That happens later (keep reading).
+//!
+//! The `Compacter` task periodically checks each relation's WAL directory to look
+//! for finished log segments, converts them to `Batch`s (basically consolidates the
+//! updates for a range of times) and adds them to the relation's `Trace`.
+//!
+//! When a `Trace` contains too many `Batches`, the Trace physically
+//! combines all of them into a single large batch with updates
+//! forwarded to the compaction frontier.
+
 use std::collections::{BTreeMap, HashMap};
 use std::fs::{self, File};
-use std::io::{Cursor, Write};
+use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::time::Duration;
 
 use anyhow::{bail, Context};
-use byteorder::{NetworkEndian, ReadBytesExt};
 use lazy_static::lazy_static;
 use log::error;
 use regex::Regex;
@@ -23,59 +58,14 @@ use tokio::sync::mpsc;
 
 use dataflow_types::Update;
 use expr::GlobalId;
-use repr::{Row, Timestamp};
+use repr::Timestamp;
 
-use crate::wal::{encode_progress, encode_update};
-
-/// Organize and maintain a compacted representation of persisted data.
-///
-/// The `Compacter` task keeps track of the persisted updates for each persisted relation and
-/// periodically compacts that representation to use space proportional to the number of
-/// distinct rows in the relation. In order to do so, the `Compacter` maintains a `Trace` for
-/// each persisted relation. Note that the `Compacter` is currently a thread running on the
-/// Materialize process, but there's no conceptual reason it couldn't be a separate process, or
-/// even on a separate machine as long as it had access to the storage for the WAL (e.g. with a
-/// shared EBS volume).
-///
-/// A `Trace` is basically a list of `Batch`s that represent a contiguous time interval, and a
-/// compaction frontier.
-///
-/// A `Batch` is a consolidated list of updates that occured between times [lower, upper)
-/// where each update is of the form `(Row, time, diff)` and each `(Row, time)` pair occurs
-/// exactly once and all diffs are nonzero.
-///
-/// Note that all `Batch`s keep their data on persistent storage. No data resides in memory
-/// (except currently we load all the data from batches into memory for compaction and on
-///  restart but this will get fixed!).
-///
-/// The `Coordinator` thread tells the `Compacter` when it needs to
-///  * start keeping track of a new relation
-///  * stop keeping track of a relation
-///  * resume keeping track of a relation with an already initialized Trace (on restart)
-///  * advance a relation's compaction frontier. Note that this doesn't automatically trigger
-///    any actual compaction. That happens later (keep reading).
-///
-/// The `Compacter` task periodically checks each relation's WAL directory to look
-/// for finished log segments, converts them to `Batch`s (basically consolidates the
-/// updates for a range of times) and adds them to the relation's `Trace`.
-///
-/// When a `Trace` contains too many `Batches`, the Trace physically
-/// combines all of them into a single large batch with updates
-/// forwarded to the compaction frontier.
+use crate::wal::{encode_progress, encode_update, read_segment, Message};
 
 // How frequently the `Compacter` checks to see if `Batch`s should be compacted.
 // TODO: Lets add some jitter to compaction so we aren't compacting every single
 // relation at the same time maybe?
 static COMPACTER_INTERVAL: Duration = Duration::from_secs(300);
-
-// Data stored in Batches and log segment files.
-#[derive(Debug, PartialEq)]
-pub enum Message {
-    // (Row, time, diff) tuples
-    Data(Update),
-    // Statements about which timestamps we might still receive data at.
-    Progress(Timestamp),
-}
 
 /// Instructions that the Coordinator sends the Compacter.
 #[derive(Debug)]
@@ -667,119 +657,4 @@ fn read_dir_regex(path: &Path, regex: &Regex) -> Result<Vec<PathBuf>, anyhow::Er
     }
 
     Ok(results)
-}
-
-/// Try to decode the next message from `buf` starting at `offset`.
-///
-/// Returns a Message and the next offset to read from if successful, or None.
-/// TODO: make this API cleaner.
-fn read_message(buf: &[u8], mut offset: usize) -> Option<(Message, usize)> {
-    if offset >= buf.len() {
-        return None;
-    }
-
-    // Let's start by only looking at the buffer at the offset.
-    let (_, data) = buf.split_at(offset);
-
-    if data.len() < 12 {
-        error!(
-            "invalid offset while reading file: {}. Expected at least 12 more bytes have {}",
-            offset,
-            data.len()
-        );
-        return None;
-    }
-
-    // Let's read the header first
-    let mut cursor = Cursor::new(&data);
-
-    let is_progress = cursor.read_u32::<NetworkEndian>().unwrap();
-
-    if is_progress != 0 {
-        // If this is a progress message let's seal a new
-        // set of updates.
-
-        // Lets figure out the time bound.
-        let timestamp = cursor.read_u64::<NetworkEndian>().unwrap();
-        // Advance the offset past what we've read.
-        offset += 12;
-
-        Some((Message::Progress(timestamp), offset))
-    } else {
-        // Let's make sure we have an appropriate number of bytes in the buffer.
-        if data.len() < 24 {
-            error!(
-                "invalid offset while reading file: {}. Expected at least 24 more bytes have {}",
-                offset,
-                data.len()
-            );
-            return None;
-        }
-        // Otherwise lets read the data.
-        let timestamp = cursor.read_u64::<NetworkEndian>().unwrap();
-        let diff = cursor.read_i64::<NetworkEndian>().unwrap() as isize;
-        let len = cursor.read_u32::<NetworkEndian>().unwrap() as usize;
-
-        assert!(diff != 0);
-
-        // Grab the next len bytes after the 24 byte length header, and turn
-        // it into a vector so that we can extract things from it as a Row.
-        // TODO: could we avoid the extra allocation here?
-        let (_, rest) = data.split_at(24);
-
-        if rest.len() < len {
-            error!(
-                "invalid row length: expected {} bytes but only have {} remaining",
-                len,
-                rest.len()
-            );
-            return None;
-        }
-        let row = rest[..len].to_vec();
-
-        let row = unsafe { Row::new(row) };
-        // Update the offset to account for the data we just read
-        offset = offset + 24 + len;
-        Some((
-            Message::Data(Update {
-                row,
-                timestamp,
-                diff,
-            }),
-            offset,
-        ))
-    }
-}
-
-/// Iterator through a set of persisted messages.
-#[derive(Debug)]
-pub struct LogSegmentIter {
-    /// Underlying data from which we read the records.
-    pub data: Vec<u8>,
-    /// Offset into the data.
-    pub offset: usize,
-}
-
-impl Iterator for LogSegmentIter {
-    type Item = Message;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some((message, next_offset)) = read_message(&self.data, self.offset) {
-            self.offset = next_offset;
-            Some(message)
-        } else {
-            None
-        }
-    }
-}
-
-impl LogSegmentIter {
-    pub fn new(data: Vec<u8>) -> Self {
-        Self { data, offset: 0 }
-    }
-}
-
-fn read_segment(path: &Path) -> Result<Vec<Message>, anyhow::Error> {
-    let data = fs::read(path).with_context(|| format!("failed to read {}", path.display()))?;
-    Ok(LogSegmentIter::new(data).collect())
 }
